# services/core/docs/openapi.yaml
openapi: 3.0.3
info:
  title: E-Commerce API
  description: |
    REST API для української e-commerce платформи.

    ## Автентифікація
    API використовує Bearer token автентифікацію. Додайте header:
    ```
    Authorization: Bearer <your-token>
    ```

    ## Рейт-ліміти
    - 100 запитів/хвилину для автентифікованих користувачів
    - 30 запитів/хвилину для неавтентифікованих

    ## Формат відповідей
    Всі відповіді повертаються в JSON форматі.

    ## Пагінація
    Використовуйте параметри `page` та `limit` для пагінації.
    Максимальний limit: 100.
  version: 1.0.0
  contact:
    name: API Support
    email: api@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: https://staging-api.example.com/v1
    description: Staging
  - url: http://localhost:8080/api/v1
    description: Development
...
paths:
  /products:
    get:
      tags:
        - Products
      summary: Список товарів
      description: Повертає список товарів з можливістю фільтрації та пагінації
      operationId: listProducts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: category_id
          in: query
          description: Фільтр по категорії
          schema:
            type: string
        - name: min_price
          in: query
          description: Мінімальна ціна
          schema:
            type: number
        - name: max_price
          in: query
          description: Максимальна ціна
          schema:
            type: number
        - name: in_stock
          in: query
          description: Тільки товари в наявності
          schema:
            type: boolean
        - name: search
          in: query
          description: Пошуковий запит
          schema:
            type: string
        - name: sort
          in: query
          description: Сортування
          schema:
            type: string
            enum: [name, -name, price, -price, created_at, -created_at]
      responses:
        '200':
          description: Успішна відповідь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - Products
      summary: Створити товар
      operationId: createProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Товар створено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /products/{id}:
    get:
      tags:
        - Products
      summary: Отримати товар
      operationId: getProduct
      parameters:
        - name: id
          in: path
          required: true
          description: ID товару
          schema:
            type: string
      responses:
        '200':
          description: Успішна відповідь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Products
      summary: Оновити товар
      operationId: updateProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Товар оновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Products
      summary: Видалити товар
      operationId: deleteProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Товар видалено
        '404':
          $ref: '#/components/responses/NotFound'
  /orders:
    get:
      tags:
        - Orders
      summary: Список замовлень
      operationId: listOrders
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Успішна відповідь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'

    post:
      tags:
        - Orders
      summary: Створити замовлення
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Замовлення створено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Отримати замовлення
      operationId: getOrder
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Успішна відповідь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/NotFound'
  /orders/{id}/status:
    patch:
      tags:
        - Orders
      summary: Оновити статус замовлення
      operationId: updateOrderStatus
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: '#/components/schemas/OrderStatus'
                tracking_number:
                  type: string
                notes:
                  type: string
      responses:
        '200':
          description: Статус оновлено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /cart:
    get:
      tags:
        - Cart
      summary: Отримати кошик
      operationId: getCart
      parameters:
        - name: cart_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Успішна відповідь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
  /cart/items:
    post:
      tags:
        - Cart
      summary: Додати товар до кошика
      operationId: addToCart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - product_id
                - quantity
              properties:
                cart_id:
                  type: string
                product_id:
                  type: string
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Товар додано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cart'
  /analytics/dashboard:
    get:
      tags:
        - Analytics
      summary: Dashboard статистика
      operationId: getDashboard
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, yesterday, week, month, year]
            default: month
      responses:
        '200':
          description: Успішна відповідь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
  /analytics/abc-xyz:
    get:
      tags:
        - Analytics
      summary: ABC-XYZ аналіз
      operationId: getAbcXyzAnalysis
      security:
        - BearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [month, quarter, year]
            default: quarter
      responses:
        '200':
          description: Успішна відповідь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbcXyzAnalysis'
  /analytics/rfm:
    get:
      tags:
        - Analytics
      summary: RFM аналіз клієнтів
      operationId: getRfmAnalysis
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Успішна відповідь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RfmAnalysis'
  /webhooks:
    get:
      tags:
        - Webhooks
      summary: Список вебхуків
      operationId: listWebhooks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Успішна відповідь
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Webhook'

    post:
      tags:
        - Webhooks
      summary: Створити вебхук
      operationId: createWebhook
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '201':
          description: Вебхук створено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'

  # Wishlist endpoints
  /wishlist/{user_id}:
    get:
      tags:
        - Wishlist
      summary: Отримати список бажань
      operationId: getWishlist
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список бажань
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WishlistItem'
    post:
      tags:
        - Wishlist
      summary: Додати в список бажань
      operationId: addToWishlist
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - product_id
              properties:
                product_id:
                  type: string
      responses:
        '201':
          description: Додано в список бажань
    delete:
      tags:
        - Wishlist
      summary: Очистити список бажань
      operationId: clearWishlist
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Список очищено
  /wishlist/{user_id}/item/{product_id}:
    get:
      tags:
        - Wishlist
      summary: Перевірити наявність товару в списку
      operationId: isInWishlist
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Статус
          content:
            application/json:
              schema:
                type: object
                properties:
                  in_wishlist:
                    type: boolean
    delete:
      tags:
        - Wishlist
      summary: Видалити зі списку бажань
      operationId: removeFromWishlist
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Видалено
  /wishlist/{user_id}/item/{product_id}/to-cart:
    post:
      tags:
        - Wishlist
      summary: Перемістити в кошик
      operationId: moveWishlistToCart
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: product_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Переміщено в кошик

  # Reviews endpoints
  /reviews:
    post:
      tags:
        - Reviews
      summary: Створити відгук
      operationId: createReview
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Відгук створено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
  /reviews/{id}:
    get:
      tags:
        - Reviews
      summary: Отримати відгук
      operationId: getReview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Відгук
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Reviews
      summary: Видалити відгук
      operationId: deleteReview
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Відгук видалено
  /products/{id}/reviews:
    get:
      tags:
        - Reviews
      summary: Відгуки на товар
      operationId: getProductReviews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Список відгуків
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
  /products/{id}/rating:
    get:
      tags:
        - Reviews
      summary: Рейтинг товару
      operationId: getProductRating
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Рейтинг
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductRating'
  /products/{id}/price-history:
    get:
      tags:
        - Products
      summary: Історія цін товару
      operationId: getPriceHistory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Історія цін
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceHistory'
  /users/{user_id}/reviews:
    get:
      tags:
        - Reviews
      summary: Відгуки користувача
      operationId: getUserReviews
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Список відгуків
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'

  # Search endpoints
  /search:
    get:
      tags:
        - Search
      summary: Пошук товарів
      operationId: searchProducts
      parameters:
        - name: q
          in: query
          description: Пошуковий запит
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: string
        - name: min_price
          in: query
          schema:
            type: number
        - name: max_price
          in: query
          schema:
            type: number
        - name: in_stock
          in: query
          schema:
            type: boolean
        - name: sort
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, name, newest]
        - $ref: '#/components/parameters/PageParam'
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Результати пошуку
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResult'
  /search/suggest:
    get:
      tags:
        - Search
      summary: Автодоповнення пошуку
      operationId: searchSuggest
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Підказки
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggestions:
                    type: array
                    items:
                      type: string
  /search/reindex:
    post:
      tags:
        - Search
      summary: Переіндексація товарів
      operationId: reindexProducts
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Переіндексацію запущено

  # Recommendations endpoints
...
  /recommendations/popular:
    get:
      tags:
        - Recommendations
      summary: Популярні товари
      operationId: getPopularProducts
      parameters:
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Популярні товари
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recommendation'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: Номер сторінки
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Кількість елементів на сторінці
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        total_pages:
          type: integer
...